backend/requirements.txt
fastapi
uvicorn
PyJWT
email-validator
pydantic[email]
python-dotenv
openai
requests